Refactor and merge RG/MMO version together. (#1)
* Removing Trailing Whitespace
* Range based for loops where reasonable.
* !GMNames.Size() replaced with GMNames.Empty()
* Remove MMOBugs specific entries.
* Remove VOID parameter from functions.
* PSPAWNINFO to PlayerClient*
* Separate INI entries to their own function. Update methods used to
obtain the information from the INI to GetPrivateProfileInt and WritePrivateProfileInt
* All WriteChatf's use same plugin message.
* Correct Typo with szGMLeaveCmdIf not being used, because it was using
szGMEnterCmdIf. The if should not work correctly.
* move globals to namespace, unsigned long to uint32_t where able. Remove
szGM[50] as it was never used.
* Add benchmark
* Rename GetValuesFromINI to SetupVolumesFromINI since there was already
a function specifically for INI changes. This one was separate for a reason I do not know, so I've opted to keep it separate from Read/Load settings functions.
* Cleanup ReadSettings() and WriteSettings() to use new functions and 
eliminate temporary variables. 
The bool values will take on a functional change and may require users to update their INI's accordingly.
* Cleanup UpdateAlerts() function.
* Removed IsGMCorpse the function is only used once.
* atoi to GetIntFromString. Change MCEval return type to int. Alter
variables to match the change for MCEval
* Replace repetitive code to alter bool values with function.
* Update GetTickCount() to MQGetTickCount64() and variables changed
to uint64_t
* Spacing, Indentation reduction, Moving variables to within scope of use
* Remove Unnecessary Variable
* use built in PluginName instead of locally created one.
* Changed PluginMsg to C string since everywhere it's used required .c_str
Suggestion by Knightly.
* Add initial logic from RG GMCheck for tracking GM history
This logic put in place as it was on the other GM Check for history purposes. WIP for alterations.
* Initial refactor of history output functions.
This compresses 3 functions down to a single function, creates an enum to determine the type of history you want to output, and makes the code a bit easier to read (imho)
* Determined the AddGMToList and RemoveGMFromList were not needed and 
added the function that writes the information to the INI when a GM is discovered. For the history.
* relocated the new logic to just above MQ plugin functions, adjusted 
the indentation on switches as per recommendation from Knightly.
* Add a TrackGMs call to the GMTest function.
* relocate TrackGMs function to above first place it's called.
* removed duplicate check found in GMCheck() function.
* Fix formatting on PluginMsg
* Not having this wrapped in () was causing it to bypass this check on
plugin load, writing all spawns in the zone to the INI. Wrapping it in () fixed that.
* Replaces the original method of getting time/date with functions present
in the MMOBugs original. The format changes to numbered dates and time, but fixes the issue where it was adding a new line at the end of the time, causing it to create a new line in the INI every time a key was updated.
* Remove debug statement.
* Remove * 25 from MAX_STRING from initial import of old code.
* Move GMHelp to just above the cmd function for ease of editing.
Add options for outputting history to the GMHelp, and Cmd, Alter bug in HistoryGMs function. Move countchars above HistoryGMs
* Remove countchars, change to PrivateProfileKeys vectors, fix GetArg use
user error where I have flipped the input and output variables. 
color and formatting for outputs
* Input TODO for INIs format. Updated current format for new date/time use
* Iterate version for major change.
* Update README.md, Remove "history" command option as it's unused. Add 
periods to end of help for zone, server, and all options to match other help entries.